[GP-689][Part 1/x]   ProofManagerFacet - L11- Code Consistency (#171)
* fix: do not use storage directly

* fix: do not use storage directly

* fix: do not use storage directly

* fix: do not isProofRevoked

* fix: build

+++ packages/greenproof-contracts/contracts/interfaces/IProofManager.sol
@@ -12,7 +12,7 @@ interface IProofManager {
 
     function getProof(uint256 certificateID) external view returns (IGreenProof.Certificate memory proof);
 
-    function getProofIdByDataHash(bytes32 dataHash) external view returns (uint proofId);
+    function getProofIdByDataHash(bytes32 dataHash) external view returns (uint256 proofId);
 
     function getProofsOf(address userAddress) external view returns (IGreenProof.Certificate[] memory);
